+++ raphtory/src/python/types/macros/iterable.rs
@@ -9,7 +9,7 @@ macro_rules! _py_numeric_methods {
             }
 
             pub fn mean(&self) -> f64 {
-                use $crate::python::types::wrappers::iterators::MeanExt;
+                use $crate::python::types::wrappers::iterables::MeanExt;
                 self.iter().mean()
             }
         }
